fix(mypy): resolve all type errors in signal_toolkit/ and remove from exclusion list
Fixed 17 mypy errors across 5 files:
- core.py: cast np.mean() returns to float for fs/dt properties
- filters.py: add wn type annotations in all 5 filter design methods
- fitting.py: annotate names dict as dict[str, Any] for mixed-type values
- noise.py: cast np.sqrt/np.std returns to float, fix floating[Any] arg
- io.py: annotate value_data dict, fix process_all return type, suppress numpy savez arg-type

Removed signal_toolkit/ from mypy exclusion list and overrides in pyproject.toml.

Closes #998 (partial - signal_toolkit directory complete)

+++ src/shared/python/signal_toolkit/core.py
@@ -62,14 +62,14 @@ class Signal:
         """Sampling frequency in Hz."""
         if len(self.time) < 2:
             return 1.0
-        return 1.0 / np.mean(np.diff(self.time))
+        return float(1.0 / np.mean(np.diff(self.time)))
 
     @property
     def dt(self) -> float:
         """Time step in seconds."""
         if len(self.time) < 2:
             return 1.0
-        return np.mean(np.diff(self.time))
+        return float(np.mean(np.diff(self.time)))
 
     @property
     def duration(self) -> float:
